# docs/openapi.yaml
...
components:
  schemas:
...
    Body_importar_api_v1_vehicles_import_post:
      properties:
        archivo:
          contentMediaType: application/octet-stream
          description: CSV exportado de Excel
          title: Archivo
          type: string
      required:
      - archivo
      title: Body_importar_api_v1_vehicles_import_post
      type: object
...
    EstadoDeVehiculo:
      description: 'Los SEIS estados — `ADR-031` cierra `IN-11`.


        `Pausado` no esta y no es un olvido: es el estado de la PUBLICACION, no del

        vehiculo. Un aviso se pausa, un auto no.'
      enum:
      - in_preparation
      - available
      - reserved
      - sold
      - in_workshop
      - archived
      title: EstadoDeVehiculo
      type: string
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    HistorialDeEstado:
      description: 'Una entrada de `GET /vehicles/{id}/history`.


        `changed_by` y `changed_at`, y NO `user_id`/`occurred_at` — `design.md`

        D-2 (C-14) los renombra contra los nombres de la TABLA (`spec-tecnica`

        §3.4, N1), que gana sobre este schema por `ADR-000`: el schema nunca tuvo

        un consumidor —ni endpoint, ni entrada en el espejo del frontend, ni en

        `test_vehiculo_espejado.py`— asi que renombrarlo no rompe nada, y mapear

        el atributo a la columna habria dejado dos vocabularios para la misma fila

        para siempre.


        `changed_by` puede ser nulo: una transicion automatica —la que dispara un

        evento de dominio, no una persona— no tiene autor.'
      properties:
        changed_at:
          format: date-time
          title: Changed At
          type: string
        changed_by:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Changed By
        from_status:
          anyOf:
          - $ref: '#/components/schemas/EstadoDeVehiculo'
          - type: 'null'
        id:
          format: uuid
          title: Id
          type: string
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
        to_status:
          $ref: '#/components/schemas/EstadoDeVehiculo'
        vehicle_id:
          format: uuid
          title: Vehicle Id
          type: string
      required:
      - id
      - vehicle_id
      - from_status
      - to_status
      - reason
      - changed_by
      - changed_at
      title: HistorialDeEstado
      type: object
...
    Transmision:
      enum:
      - manual
      - automatic
      - cvt
      - dsg
      title: Transmision
      type: string
...
    VehiculoCrear:
      additionalProperties: false
      description: 'Alta de un vehiculo.


        ⚠️ `domain_plate` es OPCIONAL — `ADR-031` cierra `IN-07`. Un 0 km sin

        patentar y un usado recien recibido en permuta no tienen dominio todavia,
        y

        con el campo obligatorio una agencia no podria cargarlos.


        Lo que si se exige es que venga **al menos uno** de dominio o chasis: un

        vehiculo sin ninguna forma de identificarse no se puede distinguir de otro.'
      properties:
        acquired_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Acquired At
        acquisition_cost_ars:
          anyOf:
          - exclusiveMinimum: 0.0
            type: number
          - pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,16}|(?=[\d.]{1,19}0*$)\d{0,16}\.\d{0,2}0*$)
            type: string
          - type: 'null'
          title: Acquisition Cost Ars
        body_type:
          $ref: '#/components/schemas/Carroceria'
        branch_id:
          format: uuid
          title: Branch Id
          type: string
        brand_id:
          format: uuid
          title: Brand Id
          type: string
        chassis_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Chassis Number
        color:
          maxLength: 60
          minLength: 1
          title: Color
          type: string
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        domain_plate:
          anyOf:
          - type: string
          - type: 'null'
          title: Domain Plate
        engine_number:
          anyOf:
          - maxLength: 30
            type: string
          - type: 'null'
          title: Engine Number
        features:
          items:
            type: string
          title: Features
          type: array
        fuel_type:
          $ref: '#/components/schemas/Combustible'
        mileage_km:
          minimum: 0.0
          title: Mileage Km
          type: integer
        model_id:
          format: uuid
          title: Model Id
          type: string
        price_ars:
          anyOf:
          - exclusiveMinimum: 0.0
            type: number
          - pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,16}|(?=[\d.]{1,19}0*$)\d{0,16}\.\d{0,2}0*$)
            type: string
          title: Price Ars
        price_usd:
          anyOf:
          - exclusiveMinimum: 0.0
            type: number
          - pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,16}|(?=[\d.]{1,19}0*$)\d{0,16}\.\d{0,2}0*$)
            type: string
          - type: 'null'
          title: Price Usd
        transmission:
          $ref: '#/components/schemas/Transmision'
        version_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Version Id
        year:
          minimum: 1950.0
          title: Year
          type: integer
      required:
      - branch_id
      - brand_id
      - model_id
      - year
      - mileage_km
      - color
      - fuel_type
      - transmission
      - body_type
      - price_ars
      title: VehiculoCrear
      type: object
...
    VehiculoSalida:
      description: 'Lo que ve CUALQUIER rol con acceso al vehiculo.


        ⚠️ `acquisition_cost_ars` NO ESTA ACA, y es el punto de que existan dos

        schemas de salida. `RN-ST-12` dice que el precio de costo solo lo ven

        `manager` y `admin_staff`, y que **nunca se publica**.


        Se resuelve con dos schemas y no con un campo opcional que el servicio

        borra: un campo que a veces viaja depende de que alguien se acuerde de

        borrarlo en cada camino de salida, y basta un `model_dump()` olvidado para

        filtrar el margen de la agencia. Con dos tipos, el que no lo tiene no puede

        filtrarlo — lo garantiza el tipo, no la disciplina.'
      properties:
        acquired_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Acquired At
        assigned_user_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Assigned User Id
        body_type:
          $ref: '#/components/schemas/Carroceria'
        branch_id:
          format: uuid
          title: Branch Id
          type: string
        brand_id:
          format: uuid
          title: Brand Id
          type: string
        chassis_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Chassis Number
        color:
          title: Color
          type: string
        created_at:
          format: date-time
          title: Created At
          type: string
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        domain_plate:
          anyOf:
          - type: string
          - type: 'null'
          title: Domain Plate
        features:
          items:
            type: string
          title: Features
          type: array
        fuel_type:
          $ref: '#/components/schemas/Combustible'
        id:
          format: uuid
          title: Id
          type: string
        mileage_km:
          title: Mileage Km
          type: integer
        model_id:
          format: uuid
          title: Model Id
          type: string
        price_ars:
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Price Ars
          type: string
        price_usd:
          anyOf:
          - pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
            type: string
          - type: 'null'
          title: Price Usd
        sold_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Sold At
        status:
          $ref: '#/components/schemas/EstadoDeVehiculo'
        tenant_id:
          format: uuid
          title: Tenant Id
          type: string
        transmission:
          $ref: '#/components/schemas/Transmision'
        version_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Version Id
        year:
          title: Year
          type: integer
      required:
      - id
      - tenant_id
      - branch_id
      - assigned_user_id
      - domain_plate
      - chassis_number
      - brand_id
      - model_id
      - version_id
      - year
      - mileage_km
      - color
      - fuel_type
      - transmission
      - body_type
      - price_ars
      - price_usd
      - status
      - description
      - features
      - acquired_at
      - sold_at
      - created_at
      title: VehiculoSalida
      type: object
info:
  description: API del SaaS multi-tenant para agencias de vehiculos.
  title: deRuedas Gestion API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/vehicles:
    get:
      description: '`T-080`, `design.md` D-2. El cuerpo sigue siendo un ARRAY de vehiculos

        —no un sobre `{items, next_cursor}`— porque

        `frontend-web/src/lib/api.ts:312` valida `Array.isArray` y un sobre lo

        rompe en tiempo de ejecucion. El cursor de la pagina siguiente viaja en el

        header `X-Next-Cursor`; su ausencia es la ultima pagina.


        `app/core/pagination.py` no se toca (C-02, CRITICO): `filtros.limit`

        llega tal cual a `acotar_tamano` adentro de `paginar`, sin revalidarlo

        acá con otro criterio — pedir mas del maximo se acota, no falla.'
      operationId: listar_vehiculos_api_v1_vehicles_get
      parameters:
      - in: query
        name: status
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/EstadoDeVehiculo'
          - type: 'null'
          title: Status
      - in: query
        name: brand_id
        required: false
        schema:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Brand Id
      - in: query
        name: model_id
        required: false
        schema:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Model Id
      - in: query
        name: branch_id
        required: false
        schema:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Branch Id
      - in: query
        name: year_from
        required: false
        schema:
          anyOf:
          - minimum: 1950
            type: integer
          - type: 'null'
          title: Year From
      - in: query
        name: year_to
        required: false
        schema:
          anyOf:
          - minimum: 1950
            type: integer
          - type: 'null'
          title: Year To
      - in: query
        name: price_from
        required: false
        schema:
          anyOf:
          - minimum: 0.0
            type: number
          - pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
            type: string
          - type: 'null'
          title: Price From
      - in: query
        name: price_to
        required: false
        schema:
          anyOf:
          - minimum: 0.0
            type: number
          - pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
            type: string
          - type: 'null'
          title: Price To
      - in: query
        name: q
        required: false
        schema:
          anyOf:
          - maxLength: 120
            minLength: 2
            type: string
          - type: 'null'
          title: Q
      - in: query
        name: cursor
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - in: query
        name: limit
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Limit
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Listar el stock
      tags:
      - stock
    post:
      description: 'Acepta `Idempotency-Key` (`T-078`): un reintento con la misma
        clave y el mismo contenido devuelve el vehiculo ya creado en vez de crear
        un segundo.'
      operationId: crear_vehiculo_api_v1_vehicles_post
      parameters:
      - in: header
        name: Idempotency-Key
        required: false
        schema:
          anyOf:
          - maxLength: 255
            minLength: 1
            type: string
          - type: 'null'
          title: Idempotency-Key
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehiculoCrear'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehiculoSalida'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Cargar un vehiculo
      tags:
      - stock
  /api/v1/vehicles/import:
    post:
      description: Acepta el CSV, lo encola y devuelve la corrida en `pending`. El
        progreso se sigue con `GET /imports/{id}`. Maximo **10 MB** y **5.000 filas**
        (`RN-ST-13`).
      operationId: importar_api_v1_vehicles_import_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_importar_api_v1_vehicles_import_post'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportacionSalida'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Importar stock desde una planilla
      tags:
      - stock
...
  /api/v1/vehicles/{vehiculo_id}:
    delete:
      description: 'Borrado logico: la fila sobrevive con `deleted_at` (regla dura
        3).'
      operationId: dar_de_baja_api_v1_vehicles__vehiculo_id__delete
      parameters:
      - in: path
        name: vehiculo_id
        required: true
        schema:
          format: uuid
          title: Vehiculo Id
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Dar de baja
      tags:
      - stock
    get:
      operationId: obtener_vehiculo_api_v1_vehicles__vehiculo_id__get
      parameters:
      - in: path
        name: vehiculo_id
        required: true
        schema:
          format: uuid
          title: Vehiculo Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Un vehiculo
      tags:
      - stock
    patch:
      description: 'Edicion PARCIAL (`PATCH`). No admite `status`: el cambio de estado
        tiene su propio endpoint porque `RN-ST-05`/`RN-ST-06` lo restringen.'
      operationId: editar_vehiculo_api_v1_vehicles__vehiculo_id__patch
      parameters:
      - in: path
        name: vehiculo_id
        required: true
        schema:
          format: uuid
          title: Vehiculo Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehiculoEditar'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Editar un vehiculo
      tags:
      - stock
  /api/v1/vehicles/{vehiculo_id}/history:
    get:
      description: 'Del cambio mas reciente al mas viejo. Mismo permiso que listar/obtener
        (`vehicles:read`), pero SIN el recorte de campos que ese permiso le aplica
        al VEHICULO (`D-4`, design.md): es una lista blanca de otro recurso, y aplicarla
        acá devolveria registros en blanco en vez de un error.'
      operationId: historial_de_vehiculo_api_v1_vehicles__vehiculo_id__history_get
      parameters:
      - in: path
        name: vehiculo_id
        required: true
        schema:
          format: uuid
          title: Vehiculo Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/HistorialDeEstado'
                title: Response Historial De Vehiculo Api V1 Vehicles  Vehiculo Id  History
                  Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: La linea de tiempo de un vehiculo
      tags:
      - stock
  /api/v1/vehicles/{vehiculo_id}/status:
    post:
      description: Solo las transiciones de `RN-ST-05`. Vender exige razon.
      operationId: cambiar_estado_api_v1_vehicles__vehiculo_id__status_post
      parameters:
      - in: path
        name: vehiculo_id
        required: true
        schema:
          format: uuid
          title: Vehiculo Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehiculoCambioDeEstado'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehiculoSalida'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Cambiar el estado
      tags:
      - stock
...